+++ ui2/src/components/QuickFilter/index.tsx
@@ -1,0 +1,3 @@
+import QuickFilter from "./QuickFilter"
+
+export default QuickFilter