+++ ui2/src/features/document-types/components/List.tsx
@@ -1,13 +1,20 @@
 import Pagination from "@/components/Pagination"
+import Th from "@/components/TableSort/Th"
 import {useGetPaginatedDocumentTypesQuery} from "@/features/document-types/apiSlice"
 import {
   clearSelection,
+  filterUpdated,
   lastPageSizeUpdate,
   selectLastPageSize,
+  selectReverseSortedByName,
   selectSelectedIds,
-  selectionAddMany
+  selectSortedByName,
+  selectTableSortColumns,
+  selectionAddMany,
+  sortByUpdated
 } from "@/features/document-types/documentTypesSlice"
-import {Center, Checkbox, Group, Loader, Stack, Table} from "@mantine/core"
+import type {DocumentTypeListColumnName} from "@/features/document-types/types"
+import {Center, Checkbox, Loader, Stack, Table} from "@mantine/core"
 import {useState} from "react"
 import {useDispatch, useSelector} from "react-redux"
 import ActionButtons from "./ActionButtons"
@@ -17,11 +24,16 @@ export default function DocumentTypesList() {
   const selectedIds = useSelector(selectSelectedIds)
   const dispatch = useDispatch()
   const lastPageSize = useSelector(selectLastPageSize)
+  const tablerSortCols = useSelector(selectTableSortColumns)
+  const sortedByName = useSelector(selectSortedByName)
+  const reverseSortedByName = useSelector(selectReverseSortedByName)
   const [page, setPage] = useState<number>(1)
   const [pageSize, setPageSize] = useState<number>(10)
   const {data, isLoading, isFetching} = useGetPaginatedDocumentTypesQuery({
     page_number: page,
-    page_size: pageSize
+    page_size: pageSize,
+    sort_by: tablerSortCols.sortBy,
+    filter: tablerSortCols.filter
   })
 
   const onCheckAll = (checked: boolean) => {
@@ -52,10 +64,27 @@ export default function DocumentTypesList() {
     }
   }
 
+  const onSortBy = (columnName: DocumentTypeListColumnName) => {
+    dispatch(sortByUpdated(columnName))
+  }
+
+  const onQuickFilterChange = (value: string) => {
+    dispatch(filterUpdated(value))
+    setPage(1)
+  }
+
+  const onQuickFilterClear = () => {
+    dispatch(filterUpdated(undefined))
+    setPage(1)
+  }
+
   if (isLoading || !data) {
     return (
       <Stack>
-        <ActionButtons />
+        <ActionButtons
+          onQuickFilterChange={onQuickFilterChange}
+          onQuickFilterClear={onQuickFilterClear}
+        />
         <Center>
           <Loader type="bars" />
         </Center>
@@ -66,7 +95,10 @@ export default function DocumentTypesList() {
   if (data.items.length == 0) {
     return (
       <div>
-        <ActionButtons />
+        <ActionButtons
+          onQuickFilterChange={onQuickFilterChange}
+          onQuickFilterClear={onQuickFilterClear}
+        />
         <Empty />
       </div>
     )
@@ -77,9 +109,12 @@ export default function DocumentTypesList() {
 
   return (
     <Stack>
-      <Group>
-        <ActionButtons /> {isFetching && <Loader size={"sm"} />}
-      </Group>
+      <ActionButtons
+        isFetching={isFetching}
+        onQuickFilterChange={onQuickFilterChange}
+        onQuickFilterClear={onQuickFilterClear}
+      />
+
       <Table>
         <Table.Thead>
           <Table.Tr>
@@ -89,7 +124,13 @@ export default function DocumentTypesList() {
                 onChange={e => onCheckAll(e.currentTarget.checked)}
               />
             </Table.Th>
-            <Table.Th>Name</Table.Th>
+            <Th
+              sorted={sortedByName}
+              reversed={reverseSortedByName}
+              onSort={() => onSortBy("name")}
+            >
+              Name
+            </Th>
           </Table.Tr>
         </Table.Thead>
         <Table.Tbody>{documentTypeRows}</Table.Tbody>