+++ changelog.md
@@ -2,6 +2,18 @@
 
 <!-- towncrier release notes start -->
 
+## [3.4] - Soon
+
+### Changes
+
+- pyproject.toml upgraded for poetry 2.0
+
+### Adds
+
+- Column sorting and filtering in Document Type list view
+- Column sorting and filtering in Custom Field list view
+
+
 ## [3.3.1] - 2025-01-19
 
 ### Changes